requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
python-dotenv>=1.0.0
persiantools>=3.0.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
pillow>=10.0.0
gunicorn>=21.0.0